+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+### Test
+- removed `AreaPrinter` test output
+- **table:** changed mock reader from `os.Stdin` to `outBuf`
+
 
 <a name="v0.12.33"></a>
 ## [v0.12.33] - 2021-10-24